requirements.txt
asttokens==2.0.5
backcall==0.2.0
click==8.1.3
colorama==0.4.5
customtkinter==4.5.2
darkdetect==0.6.0
debugpy==1.6.0
decorator==5.1.1
dnspython==2.2.1
email-validator==1.2.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
Flask==2.2.1
Flask-Login==0.6.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
ipykernel==6.15.0
ipython==8.4.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jupyter-client==7.3.4
jupyter-core==4.10.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.5
numpy==1.23.1
openpyxl==3.0.10
packaging==21.3
pandas==1.4.3
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.30
psutil==5.9.1
psycopg2==2.9.3
pure-eval==0.2.2
Pygments==2.12.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
pyzmq==23.2.0
six==1.16.0
SQLAlchemy==1.4.39
stack-data==0.3.0
tornado==6.1
traitlets==5.3.0
wcwidth==0.2.5
Werkzeug==2.2.1
WTForms==3.0.1
